### 2019-07-08 #1549 @labarba
Thank you, @rouson!! You can find the [Reviewer Guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) on our dedicated documentation site.

Do you know anyone who could be a good second reviewer?
