### 2019-07-08 #1549 @labarba
👋  @hariradh, @zbeekman — Would either of you be willing to provide a review for a submission to the Journal of Open Source Software?
*FEST-3D: Finite-volume Explicit STructured 3-Dimensional solver*
software: https://github.com/FEST3D/FEST-3D
paper tmp link: https://github.com/openjournals/joss-papers/blob/joss.01549/joss.01549/10.21105.joss.01549.pdf

[Reviewer Guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) on our dedicated documentation site 🚀
