### 2019-07-09 #1549 @hariradh
Dear Lorena,

I am willing to review the submission. Please send me any additional information needed for the review.

Regards,
Hari

Sent from my iPhone

> On Jul 8, 2019, at 11:07 PM, Lorena A. Barba <notifications@github.com> wrote:
>
> 👋 @hariradh, @zbeekman — Would either of you be willing to provide a review for a submission to the Journal of Open Source Software?
> FEST-3D: Finite-volume Explicit STructured 3-Dimensional solver
> software: https://github.com/FEST3D/FEST-3D
> paper tmp link: https://github.com/openjournals/joss-papers/blob/joss.01549/joss.01549/10.21105.joss.01549.pdf
>
> Reviewer Guidelines on our dedicated documentation site 🚀
>
