### 2019-07-09 #1549 @labarba
That's great @hariradh — Thank you. I will add you and start the review now. This will open a new GitHub issue, with a review checklist for you. JOSS reviews happen in the open, with reviewer comments posted on the review issue, and you can also open issues on the software repository. I'm here to answer any questions!
