### 2019-07-09 #1555 @labarba
👋 @hariradh, @rouson — This is where the action happens. You each have a review checklist above to work through. As you do, feel free to post your reviewer comments to the author in this thread. You may also open issues in the software repository as needed, and post a link here. Ask me anything!

[Reviewer Guidelines](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) on our dedicated documentation site 🚀
