### 2019-07-23 #1555 @hariradh
I have trouble compiling the code based on the instructions. The `cmake` step hangs.
```
$ FC=mpif90 cmake ..
-- The Fortran compiler identification is GNU 9.1.0
-- Checking whether Fortran compiler has -isysroot
-- Checking whether Fortran compiler has -isysroot - yes
-- Checking whether Fortran compiler supports OSX deployment target flag
-- Checking whether Fortran compiler supports OSX deployment target flag - yes
-- Check for working Fortran compiler: /usr/local/bin/mpif90
-- Check for working Fortran compiler: /usr/local/bin/mpif90  -- works
-- Detecting Fortran compiler ABI info
-- Detecting Fortran compiler ABI info - done
-- Checking whether /usr/local/bin/mpif90 supports Fortran 90
-- Checking whether /usr/local/bin/mpif90 supports Fortran 90 -- yes
-- CMAKE_BUILD_TYPE not given, defaulting to RELEASE
```

I am testing the code using:
- Cmake 3.15
- gcc 9.1.0
- MPICH2 3.3.1
- OSX 10.14.5

`Cmake`, `gcc`,  and `MPICH2` were installed using `homebrew`.
