### 2019-07-24 1288532
Fixed the CMake Module for MacOSX
The -fno-underscoring flag creates a problem for the UNIX operating system (MacOSX ). It is hence removed.
