### 2019-07-24 #1555 @jayten
The compilation problem on MacOSX has been sorted out. The following piece of code in the CMake module "SetFortranFlag.cmake" was creating a problem.
```
# Don't add underscores in symbols for C-compatibility
SET_COMPILE_FLAG(CMAKE_Fortran_FLAGS "${CMAKE_Fortran_FLAGS}"
                 Fortran "-fno-underscoring")
```
So, the issue was corrected on [**Fixed the Compatibility issue of MacOSX for CMake script**](https://github.com/FEST3D/FEST-3D/pull/4) branch and merged with master.

The latest state of the code on master branch has been test on OSX version 10.13.4 using following:
- CMake-3.15.0
-  Gcc-9.1.0
-  Libmpc-1.1.0
-  Mpfr-4.0.2
-  Isl-0.21
-  Gmp-6.1.2_2
-  mpich-3.3.1

Compilation was done using
 `$ FC=mpif90 cmake ..`
and then
` $ make -j 4`
