### 2019-08-16 #1555 @hariradh
@labarba @jayten

I apologize for the delay but I am having trouble testing the functionality of the code. The code fails when running the `automaton.py` script in the `CheckInput` function. I was trying to figure out the cause before sending my comments to the authors but have not been able to complete the task due to other priorities. Here are a couple of examples:
```
ginger:run hari$ python automaton.py
Traceback (most recent call last):
  File "automaton.py", line 347, in <module>
    CheckInput(ExpectedControl, ExpectedScheme, ExpectedFlow, ExpectedOutputControl, ExpectedResidualControl, Control, Scheme, Flow, OutputControl, ResidualControl)
  File "automaton.py", line 187, in CheckInput
    assert all(variable in ExpectedOutputControl['In'] for variable in OutputControl['In'])
AssertionError
```
```
ginger:LidDrivenCavity hari$ python automaton.py
Traceback (most recent call last):
  File "automaton.py", line 340, in <module>
    CheckInput(ExpectedControl, ExpectedScheme, ExpectedFlow, ExpectedOutputControl, ExpectedResidualControl, Control, Scheme, Flow, OutputControl, ResidualControl)
  File "automaton.py", line 183, in CheckInput
    assert len(next(os.walk(GridDir))[2]) == NumberOfBlocks
StopIteration
```
