### 2019-08-16 #1555 @jayten
### First error:
The **automaton.py**  in the run directory is a sample script file. This script file is not supposed to be executed as the inputs are randomly provided as an explanation for the variables. Users are supposed to copy this script and modify as according to need. In the **Tutorial** subfolder, different **automation.py** file is provided with specific inputs.

### Second error:
The assertion error indicates that the number of blocks provided as input in the file (**NumberOfBlocks**) does not match with the number of grid files available in the **GridDir**. So,  grid files are not created before calling setup script **automaton.py**. This assertion helps to avoid the breakdown of the solver after it is executed. Each block in the solver requires an individual grid file.

As described in the [documentation](https://fest3d.github.io/page/05_tutorials/03_LDC.html), the grid files for Lid-driven cavity test case can be created easily by just switching to the **CreatedBlock** subfolder and executing **make** command. An important point to note here is that the grid generation requires  **numpy** version 1.10 or above for **Python2.7**. Once the grids the generated, the error should not occur.

Try following steps in the LidDrivenCavity folder:
```
$ cd CreateBlocks/
$ make
```
An error might occur if "numpy" version is less than 1.10 as the function "stack" is not implemented in the older versions.
```
$ cd ../
$ python automaton.py
```

Another commonly occurring mistake is not defining the path of the **AbsBinaryPath**. All the other input in the automaton.py will work without changing, but **AbsBinaryPath** is the absolute path to the directory where FEST-3D binary is compiled on your local computer.

In the **automaton.py** file make sure you change following:
AbsBinaryPath="/Path/to/FEST-3D/bin/FEST3D"

More details can be found in the [Documentation](https://fest3d.github.io/page/index.html)
