### 2019-09-11 #1555 @Kevin-Mattheus-Moerman
@rouson thanks again for helping us review this JOSS submission. Can you please update us on your progress with this review? If you need more time please inform us as well. Thanks!
