+++ test/geometry.f90
@@ -1,0 +1,261 @@
+  subroutine compute_i_face_area(imx, jmx, kmx, grid_x, grid_y, grid_z, xn)
+      !-----------------------------------------------------------
+      ! Compute face area vectors
+      !
+      ! The face area vectors denote the face area both in 
+      ! magnitude and direction. They are placed in the face 
+      ! normal variables for further calculation.
+      !
+      ! The face normal is given by d1 x d2, where d1 and d2 are
+      ! the diagonals of a face
+      !-----------------------------------------------------------
+      
+      implicit none
+      integer , intent(in) :: imx, jmx, kmx
+      real*8 , dimension(1:imx, 1:jmx, 1:kmx), intent(in) :: grid_x, grid_y, grid_z
+      real*8 , dimension(1:imx, 1:jmx-1, 1:kmx-1,1:3), intent(out) :: xn
+
+      real*8 :: d1x, d2x, d1y, d2y, d1z, d2z
+      integer :: i, j, k
+
+      !f2py intent(in) :: imx, jmx, kmx
+      !f2py intent(in) :: grid_x, grid_y, grid_z
+      !f2py intent(out) :: xn
+
+      do k = 1, kmx-1
+       do j = 1, jmx-1
+        do i = 1, imx
+          d1x = grid_x(i, j+1, k+1) - grid_x(i, j, k)
+          d1y = grid_y(i, j+1, k+1) - grid_y(i, j, k)
+          d1z = grid_z(i, j+1, k+1) - grid_z(i, j, k)
+          d2x = grid_x(i, j, k+1) - grid_x(i, j+1, k)
+          d2y = grid_y(i, j, k+1) - grid_y(i, j+1, k)
+          d2z = grid_z(i, j, k+1) - grid_z(i, j+1, k)
+          xn(i, j, k,1) = 0.5 * (d1y*d2z - d1z*d2y)
+          xn(i, j, k,2) = 0.5 * (d1z*d2x - d1x*d2z)
+          xn(i, j, k,3) = 0.5 * (d1x*d2y - d1y*d2x)
+         end do
+        end do
+       end do
+
+  end subroutine compute_i_face_area
+    
+
+  subroutine compute_j_face_area(imx, jmx, kmx, grid_x, grid_y, grid_z, yn)
+      implicit none
+      integer , intent(in) :: imx, jmx, kmx
+      real*8 , dimension(1:imx, 1:jmx, 1:kmx), intent(in) :: grid_x, grid_y, grid_z
+      real*8 , dimension(1:imx-1, 1:jmx, 1:kmx-1,1:3), intent(out) :: yn
+
+      real*8 :: d1x, d2x, d1y, d2y, d1z, d2z
+      integer :: i, j, k
+
+      !f2py intent(in) :: imx, jmx, kmx
+      !f2py intent(in) :: grid_x, grid_y, grid_z
+      !f2py intent(out) :: yn
+
+      do k = 1, kmx-1
+       do j = 1, jmx
+        do i = 1, imx-1
+         d1x = grid_x(i+1, j, k+1) - grid_x(i, j, k)
+         d1y = grid_y(i+1, j, k+1) - grid_y(i, j, k)
+         d1z = grid_z(i+1, j, k+1) - grid_z(i, j, k)
+         d2x = grid_x(i+1, j, k) - grid_x(i, j, k+1)
+         d2y = grid_y(i+1, j, k) - grid_y(i, j, k+1)
+         d2z = grid_z(i+1, j, k) - grid_z(i, j, k+1)
+      
+         yn(i, j, k,1) = 0.5 * (d1y*d2z - d1z*d2y)
[…]
