### 2019-09-30 #1555 @jayten
@hariradh

### API Documentation
The [FEST-3D website](https://fest3d.github.io/index.html) provides detailed [documentation](https://fest3d.github.io/page/index.html) of different [source files](https://fest3d.github.io/lists/files.html), [modules](https://fest3d.github.io/lists/modules.html), [procedures](https://fest3d.github.io/lists/procedures.html), and [main program](https://fest3d.github.io/program/main.html). This documentation is generated using FORD documentation generator. You can follow the links to find information on all the subroutines and functions written in different modules.

### Unit Tests
It is rather difficult to write unit tests for the [FEST-3D code](https://github.com/FEST3D/FEST-3D); therefore, we have written integrated tests instead. FEST-3D repository has been updated with a new folder: test/. Please pull the changes in the **master** branch on GitHub. A python script, "**Test.py**," provided in the **test/** directory to run the integrated tests with a particular flux scheme, higher-order method, and turbulence model. Use the following command to run all tests in the **test/** directory. After generating the binary, you can run the test script using the following command in the **test/** folder:
```
$cd <FEST-3D/root/folder>/test/
$python Test.py <arg1>  <arg2>  <arg3>
```

1.  <arg1> : Flux scheme
    - allowed options:
         - ausm
        - ldfss0
        - ausmUP
        - ausmP
        - slau

2. <arg2> : Higher-order method
    - allowed options:
        - muscl
        - ppm
        - weno

3. <arg3> : Turbulence model
    - allowed options:
        - sst
        - sst2003
        - kkl
        - sa

Examples:
 1. `$python Test.py ausm muscl sst`
 2. `$python Test.py slau weno sa`
 3. `$python Test.py ausmUP ppm kkl `

Please note that any flux-reconstruction scheme/higher-order method/turbulence model added by a developer can also be tested using **Test.py** script by suitably modifying the arguments passed to the script.

Once the tests are complete, you can check the summary in the **Report.txt** file. For more details about the integrated tests, please go through the [TEST](https://fest3d.github.io/page/03_test.html) webpage in the [FEST-3D documentation](https://fest3d.github.io/page/index.html).

### How to run FEST-3D
Also, as request by the @hariradh, we have also updated the [website](https://fest3d.github.io/page/index.html) with a new page on "[How to run FEST-3D](https://fest3d.github.io/page/04_Steps_to_run_FEST3D.html)."  This page provides a step-by-step approach on [how to run the Lid-driven cavity test case](https://fest3d.github.io/page/04_Steps_to_run_FEST3D.html#steps-to-LDC).
