### 2019-10-01 #1555 @rouson
@labarba Yes, I believe I can complete my review within a week of today.
