### 2019-10-05 #1555 @labarba
You mean, while logged in to GitHub, you click on the box, and nothing happens? It could be a permissions issue. Though you should have been automatically added to the right team when added as a reviewer. @arfon : could you check this?
