### 2019-11-17 #1555 @rouson
@labarba thanks for the email.  I somehow got subscribed to every JOSS communication so these messages got buried in the resulting flood.  Hopefully I just resolved it by clicking "Unwatch."  I'll work on my review later today and post the results.
