### 2019-12-01 #1555 @rouson
@labarba I have added two issues that need to be addressed in order for me to complete my review.  In addition, I am concerned about the appropriateness of releasing new software based on a language version that is very nearly three decades old (Fortran 90) and using a parallel programming model that is over two decades old (MPI-1).  The Fortran standard has undergone 4 revisions since Fortran 90, each of which has added features that could be used to make the authors' code considerably more clear and concise and could potentially have benefits in robustness and performance.  MPI has undergone two revisions since MPI-1 and more importantly, the most recent Fortran standards provide a sufficiently rich set of parallel features that it's not even clear there is a need for embedding MPI directly in _new_ projects.  The compiler and/or parallel runtime library can generate the MPI and would likely generate more sophisticated MPI (or an alternative parallel programming model) than scientific application developers would likely write so again there's the possibility for greater clarity, conciseness, robustness, and performance.
