### 2019-12-01 FEST3D/FEST-3D#7 @rouson
I updated the comment above to eliminate several duplicate items that were caused by an inadvertent copy and paste error.
