### 2019-12-01 #1555 @rouson
@labarba @danielskatz  I just added [another issue] on this submissions repository with an 10-item list of modernization suggestions.  Despite the list's length, it is surely not exhaustive as I've only reviewed a few files.  I'd like your guidance on whether archaic use of the language is a relevant review criterion.  I can't see where it fits into any of the items in my reviewer checklist, but after writing this latest issue, I would recommend rejection unless you're of the opinion that my concern is not part of the allowable criteria.  I feel strongly that it's time for the world to move on from coding to a language standard that is very nearly 3 decades old.  The code could definitely be more concise and robust and might even be higher performing if it were extensively overhauled and modernized to exploit the Fortran 2018 features that are supported by the compiler recommended in the installation instructions (`gfortran`).  In a much more modern form, the code could be a great contribution to the literature.

[another issue]: https://github.com/FEST3D/FEST-3D/issues/7
