### 2019-12-01 FEST3D/FEST-3D#7 @rouson
Other Fortran intrinsic functions that will likely greatly reduce the side of your code are `count` and `sum`.  Please familiarize yourself with these and use them liberally unless you have a very strong motivation for doing otherwise such as solid performance data showing they slow down your application.  For example, [subroutine apply_interface] appears to be not much more than hundreds of lines of code blocks comprised of loops nested 4 levels deep for no purpose other than computing a sum that could presumably be computed in a single line of code by invoking the `sum` function. You could achieve an order-or-of-magnitude reduction in code length for this subroutine alone. The change has important ramifications:

1. **Code clarity:** readers take more time to figure out what you're doing when they have to read 10 lines of code versus 1 -- especially because the reader will likely give you the benefit of the doubt and assume there is some good reason you chose the more verbose approach and will spend time trying to figure out the reason (I can't find one in this case).
2. **Speed:**  As mentioned elsewhere, loops are explicitly serial by definition.  At high levels of optimization, a smart compiler might recognize the pattern in your code and replace it with faster code (likely by invoking `sum`), but it's generally better to give the compiler a high-level description of what you want (an arithmetic sum) and let the compiler choose the fastest way, which often could involve vectorization in this case and vectorization is a form of parallelism.

People often talk about how to parallelize an algorithm.  I frequently point out that so much of nature and even so much of what we're doing when programming is inherently parallel so the more important question is "Why serialize the algorithm?"

[subroutine apply_interface]: https://github.com/FEST3D/FEST-3D/blob/master/src/plusgs.f90#L2885
