### 2019-12-01 #1555 @rouson
I just added yet another [issue] as well as a lengthy comment on a previous issue.  The more I look at the code, the more confident I am in recommending rejection.  I'm certain this code addresses an important problem and is therefore useful, but it really needs a major revision to come up to the standards of what I would expect for any modern open-source package.

[issue]: https://github.com/FEST3D/FEST-3D/issues/7
