+++ src/face/state/weno.f90
@@ -14,9 +14,6 @@ module weno
     use global_vars, only : imx
     use global_vars, only : jmx
     use global_vars, only : kmx
-    use global_vars, only : grid_x
-    use global_vars, only : grid_y
-    use global_vars, only : grid_z
 
     use global_vars, only : qp
     use global_vars, only : n_var
