### 2019-12-03 FEST3D/FEST-3D#7 @rouson
### Regarding Fortran standards
Each Fortran standard is very nearly backwards compatible.  Very few features get deleted or modified so Fortran 90 is very nearly subset of Fortran 2018 and I don't see any obvious cases where you're using features that have been deleted from the language.  In this sense, you're already writing Fortran 2018.  The key question is why restrict yourself to a subset of the language that is nearly 30 years old?  Unless the Fortran committee has wasted the last 3 decades, there almost has to be features of value that you miss by by such a severe restriction. Based on my review so far, I believe there are many and I suspect my list would grow if I reviewed more of the code.  I also think it does the Fortran community a disservice to write new code in such an old subset of the language.  For one thing, it limits the community of people who will be interested in contributing to the project.  When I see projects writing new code in Fortran 90, I ask how will you attract developers under the age of 30 when you tell them they can't use any language features that were invented in their entire lifetime?  That can be a hard sell.

### Regarding performance and/or memory usage
I never believe performance claims or memory utilization claims unless they come with specific empirical data about the subset of code being discussed and its impact on the overall application performance.  Before asking someone to read through hundreds of lines of loops nested 4 levels deep that could be made roughly 90% shorter by simply invoking the `sum` intrinsic function or another intrinsic function, I would want to see a statement such as "This specific block of code occupies x% of the overall application runtime in important use cases y and is sped up by z% by writing nested loops instead of simply calling `sum`."

### Regarding compiler versions
* With the exception of the `mpi_f08` module, every feature I named should be supported by multiple earlier versions of `gfortran`.  Lack of access to version 9 is not a valid reason to write code to a 30-year-old standard.  In fact, in most of the ways I mentioned, you're not even taking sufficient advantage of Fortran 90's level of expressiveness.  The ability to avoid verbose, handwritten manipulations of arrays is one of the main benefits of Fortran 90 and only gets better with subsequent standards because of all the intrinsic functions that have been added.
* Also, I would be glad to advise you on how to install `gfortran` version 9.  One way would be to use the OpenCoarrays installer, which I wrote.  Doing so should take no more than three commands:
```bash
git clone https://github.com/sourceryinstitue/opencoarrays
cd opencoarrays
./install.sh -p gcc -I 9.2.0 -i <desired-installation-path> -j 4
```

### Regarding MPI
I'm not an MPI expert so I'll defer to your knowledge about how non-blocking communication has performed in your application.  An even more important point is that it's not clear to me why _new_ software would embed MPI of any form in the source code when Fortran has had its own parallel programming abstractions since Fortran 2008.  If you go the above route to install `gfortran`, you can also use the same installer to install OpenCoarrays, which then gives you access to parallel features that completely obviate the need to embed MPI in your source.  The parallel runtime library can generate the MPI commands for you.  Most importantly, it will generate MPI-3 one-sided communication which is more advanced than even the aforementioned non-blocking communication.  This might or might not speed up your application, but it's certainly a good place to start when writing _new_ code.  Moreover, it opens up the possibility to swap MPI for another parallel programing model such as OpenSHMEM without changing even one line of source code.

I wouldn't write any of what I'm writing if this were an old project building on decades-old legacy code.  I just don't think new code should be written in the manner I see in this project.
