### 2019-12-03 #1555 @labarba
@rouson --- Thank you very much for the detailed recommendations for improvements that you have provided to the authors.

Note, though, that the JOSS review process seldom leads to a _rejection_, in the manner of traditional journals. The Review Guidelines, in the section on [**What happens if the software I’m reviewing doesn’t meet the JOSS criteria?**](https://joss.readthedocs.io/en/latest/review_criteria.html#what-happens-if-the-software-i-m-reviewing-doesn-t-meet-the-joss-criteria) says:

> We ask that reviewers grade submissions in one of three categories: 1) Accept 2) Minor Revisions 3) Major Revisions. Unlike some journals we do not reject outright submissions requiring major revisions - we’re more than happy to give the author as long as they need to make these modifications/improvements.

In other words, we aim to help authors improve their software _until it passes the bar_ to be accepted. They can, of course, decide that they are now willing to do the improvements, and request to withdraw the submission.

In this particular case, your critique of the software will need to be assessed by the authors, and they are allowed to come back to us with a proposal of what / how much they want to implement of your suggestions. The authors are **not required** to use particular languages or language features. JOSS reviewers/editors may insist for updates in the case of languages no longer supported (or soon to be), like Python 2.x. That is not the case with old Fortran, which is very much still used and supported.

It is the case, however, that using old languages may make the software. less likely to be reused, less likely for others to contribute to it, and less likely to be cited. This is not grounds for rejection, but should be of keen interest to the authors.

I encourage the authors to assess the recommendations, and the improvements they would like to make. For example, this critique:

> “hundreds of lines of code blocks comprised of loops nested 4 levels deep for no purpose other than computing a sum that could presumably be computed in a single line of code by invoking the `sum` function.”

... it sounds like a sizable reduction in number of lines of code is possible with some modernizing. Yet, the authors could have concerns with compilers available in their machines supporting the modern features.

@jayten — Please discuss with your team, and get back to us with your assessment.
