### 2019-12-05 #1555 @rouson
@labarba I love that and fully support everything you wrote.  JOSS is a breath of fresh air!
