### 2019-12-07 #1555 @jayten
We are thankful to the reviewer @rouson for his insightful comments. Our team is not familiar with the latest Fortran version, but based on the reviewer's suggestion, it seems like the right way forward to use newer features supported by the latest Fortran version.  However, at the same time, we want to make sure that It does not create any conflict with our already planned work - implementation of  OpenACC directive based parallelisation of loops. As Gfortran9 only provides [experimental and incomplete](https://gcc.gnu.org/onlinedocs/gfortran/OpenACC.html) support for OpenACC, the new members of our team have chosen the PGI compiler to write directive-based GPU programming. As such, Coarray seems to be very simple and productive to use, the PGI compiler appears to be holding back in supporting Coarray as they are still only [FORTRAN 2003 compliant](http://fortranwiki.org/fortran/show/Compiler+Support+for+Modern+Fortran). We are also not sure if array operation is compatible with OpenACC as it seems from our preliminary reading that OpenACC directives recognise the loops (which are to be run parallel in GPU) from the "do" and "end do" statements. We are still new to this field (use of OpenACC directives) and will be happy to be corrected if we are wrong.

Since most of the other suggestions by the reviewer are doable without any conflict with our future work, we propose to do the following work for now:

1. Reducing the length of the code wherever possible (for eg: a single subroutine for all three directions of the 3D grid, use of derived datatype, etc. )
2.  Making variables local to the module.
3.  Instead of using just global variables, passing arguments to subroutines.
4.  Write unit tests with ctest.
5.  Remove unnecessary comments and files from the code.

**A point to note**:  the portion of code that the reviewer suggested could be replaced with just a sum statement is actually a buffer collection operation, wherein it is then passed across nodes using MPICH MPISend and MPIRecv commands. We realise though that with the use of Coarrays this will be just a few lines of code.
