### 2019-12-14 de02e96
Add: controltype for control file
