+++ src/CC.f90
@@ -8,9 +8,10 @@ module CC
 #include "debug.h"
 #include "error.h"
 
-  use global_vars, only: imx
-  use global_vars, only: jmx
-  use global_vars, only: kmx
+   use vartypes
+!  use global_vars, only: imx
+!  use global_vars, only: jmx
+!  use global_vars, only: kmx
   use global_vars, only: process_id
   use global_vars, only: xn
   use global_vars, only: yn
@@ -43,21 +44,22 @@ module CC
   
   contains
 
-    subroutine setupCC()
+    subroutine setupCC(dims)
       !< Allocate memory for the cell center variable only in case of transition model
       implicit none
+      type(extent), intent(in) :: dims
 
       DebugCall("Setup CC")
 
       if((transition=='lctm2015') .and. turbulence/='none')then
-        call alloc(CCnormalX, -2, imx+2, -2, jmx+2, -2, kmx+2, AErrMsg("CCnormalX"))
-        call alloc(CCnormalY, -2, imx+2, -2, jmx+2, -2, kmx+2, AErrMsg("CCnormalY"))
-        call alloc(CCnormalZ, -2, imx+2, -2, jmx+2, -2, kmx+2, AErrMsg("CCnormalZ"))
-        call alloc(CCVn,   -2, imx+2, -2, jmx+2, -2, kmx+2, AErrMsg("CCVn"))
-        call alloc(DCCVnX, -2, imx+2, -2, jmx+2, -2, kmx+2, AErrMsg("DCCVnZ"))
-        call alloc(DCCVnY, -2, imx+2, -2, jmx+2, -2, kmx+2, AErrMsg("DCCVnY"))
-        call alloc(DCCVnZ, -2, imx+2, -2, jmx+2, -2, kmx+2, AErrMsg("DCCVnZ"))
-        call find_CCnormal()
+        call alloc(CCnormalX, -2, dims%imx+2, -2, dims%jmx+2, -2, dims%kmx+2, AErrMsg("CCnormalX"))
+        call alloc(CCnormalY, -2, dims%imx+2, -2, dims%jmx+2, -2, dims%kmx+2, AErrMsg("CCnormalY"))
+        call alloc(CCnormalZ, -2, dims%imx+2, -2, dims%jmx+2, -2, dims%kmx+2, AErrMsg("CCnormalZ"))
+        call alloc(CCVn,   -2, dims%imx+2, -2, dims%jmx+2, -2, dims%kmx+2, AErrMsg("CCVn"))
+        call alloc(DCCVnX, -2, dims%imx+2, -2, dims%jmx+2, -2, dims%kmx+2, AErrMsg("DCCVnZ"))
+        call alloc(DCCVnY, -2, dims%imx+2, -2, dims%jmx+2, -2, dims%kmx+2, AErrMsg("DCCVnY"))
+        call alloc(DCCVnZ, -2, dims%imx+2, -2, dims%jmx+2, -2, dims%kmx+2, AErrMsg("DCCVnZ"))
+        call find_CCnormal(dims)
       end if
 
     end subroutine setupCC
@@ -80,12 +82,13 @@ module CC
     end subroutine destroyCC
 
 
-    subroutine find_CCnormal()
+    subroutine find_CCnormal(dims)
       !< Find the cell-center unit normal
       implicit none
-      call compute_gradient(CCnormalX, dist, 'x')
-      call compute_gradient(CCnormalY, dist, 'y')
-      call compute_gradient(CCnormalZ, dist, 'z')
+      type(extent), intent(in) :: dims
[…]
