### 2019-12-16 9b9f640
ADD: schemetype and flowtype
