+++ src/read_write/write/write_output_tec.f90
@@ -11,12 +11,6 @@ module write_output_tec
   use global     , only : outin_file
   use vartypes
 
-!  use global_vars, only : write_data_format
-!  use global_vars, only : write_file_format
-!  use global_vars, only : imx
-!  use global_vars, only : jmx
-!  use global_vars, only : kmx
-  !use grid, only : point
   use global_vars, only : density 
   use global_vars, only : x_speed 
   use global_vars, only : y_speed 
@@ -29,7 +23,6 @@ module write_output_tec
   use global_vars, only : tgm
   use global_vars, only : mu 
   use global_vars, only : mu_t 
-  use global_vars, only : gm
   use global_vars, only : dist
   use global_vars, only : vis_resnorm
   use global_vars, only : cont_resnorm
@@ -48,12 +41,8 @@ module write_output_tec
   use global_vars, only : tv_residue
   use global_vars, only : intermittency
 
-  use global_vars, only : turbulence
-  use global_vars, only : mu_ref
-!  use global_vars, only : current_iter
-!  use global_vars, only : max_iters
-  use global_vars, only : w_count
-  use global_vars, only : w_list
+  !use global_vars, only : w_count
+  !use global_vars, only : w_list
 
   use global_sst , only : sst_F1
   use gradients, only : gradu_x
@@ -75,10 +64,8 @@ module write_output_tec
   use gradients, only : gradtw_y
   use gradients, only : gradtw_z
   use global_vars, only : process_id
-!  use global_vars, only : checkpoint_iter_count
 
   use utils
-!  use string
 
   implicit none
   private
@@ -89,12 +76,12 @@ module write_output_tec
 
   contains
 
-    subroutine write_file(nodes, dims, checkpoint_iter_count)
+    subroutine write_file(nodes, control, dims)
       !< Write the header and variables in the file "process_xx.dat".
       implicit none
+      type(controltype), intent(in) :: control
       type(extent), intent(in) :: dims
       type(nodetype), dimension(-2:dims%imx+3,-2:dims%jmx+3,-2:dims%kmx+3), intent(in) :: nodes 
-      integer, intent(in) :: checkpoint_iter_count
       integer :: n
       character(len=*), parameter :: err="Write error: Asked to write non-existing variable- "
 
@@ -102,12 +89,12 @@ module write_output_tec
       imx = dims%imx
       jmx = dims%jmx
       kmx = dims%kmx
-      call write_header(checkpoint_iter_count)
+      call write_header(control)
       call write_grid(nodes)
 
-      do n = 1,w_count
+      do n = 1,control%w_count
 
-        select case (trim(w_list(n)))
+        select case (trim(control%w_list(n)))
         
           case('Velocity')
             call write_scalar(x_speed, "u", -2)
@@ -233,7 +220,7 @@ module write_output_tec
             continue
 
           case Default
-            print*, err//trim(w_list(n))//" to file"
+            print*, err//trim(control%w_list(n))//" to file"
 
         end select
       end do
@@ -242,10 +229,10 @@ module write_output_tec
     end subroutine write_file
 
 
-    subroutine write_header(checkpoint_iter_count)
+    subroutine write_header(control)
       !< Write the header in the output file in the tecplot format
       implicit none
-      integer, intent(in) :: checkpoint_iter_count
+      type(controltype), intent(in) :: control
       integer :: n
       integer :: total
 
@@ -253,9 +240,9 @@ module write_output_tec
       write(OUT_FILE_UNIT,'(a)') "variables = x y z "
 
[…]
