### 2019-12-19 d536e31
Add: read and write variable to controltype
