### 2019-12-19 cc7d984
Switched: global state variables to local
