+++ src/update.f90
@@ -28,35 +28,28 @@ module update
   use gradients  ,only :   gradw_y
   use gradients  ,only :   gradw_z
   use global_vars, only : volume
-!  use global_vars, only : qp
-!  use global_vars, only : density
-!  use global_vars, only : x_speed
-!  use global_vars, only : y_speed
-!  use global_vars, only : z_speed
-!  use global_vars, only : pressure
   use global_vars, only : dist
   use global_vars, only : mu
   use global_vars, only : mu_t
   use global_vars, only : delta_t
   use global_vars, only : process_id
 
-  use global_vars, only: F_p
-  use global_vars, only: G_p
-  use global_vars, only: H_p
-  use global_vars, only: mass_residue
-  use global_vars, only: x_mom_residue
-  use global_vars, only: y_mom_residue
-  use global_vars, only: z_mom_residue
-  use global_vars, only: energy_residue
-  use global_vars, only: TKE_residue
-  use global_vars, only: omega_residue
-  use global_vars, only: kl_residue
-  use global_vars, only: residue
+!  use global_vars, only: F_p
+!  use global_vars, only: G_p
+!  use global_vars, only: H_p
+!  use global_vars, only: mass_residue
+!  use global_vars, only: x_mom_residue
+!  use global_vars, only: y_mom_residue
+!  use global_vars, only: z_mom_residue
+!  use global_vars, only: energy_residue
+!  use global_vars, only: TKE_residue
+!  use global_vars, only: omega_residue
+!  use global_vars, only: kl_residue
+!  use global_vars, only: residue
 
   use geometry   , only: CellCenter
 
   use utils, only: alloc
-  use utils, only:  dealloc 
 
   !subroutine for residual calculation
   use interface1,                      only: apply_interface
@@ -180,7 +173,7 @@ module update
 !      end subroutine destroy_update
 
 
-      subroutine get_next_solution(qp, Temp, control, scheme, flow, dims)
+      subroutine get_next_solution(qp, Temp, residue, F,G,H, control, scheme, flow, dims)
         !< Get solution at next time-step using scheme
         !< given in the input file.
         implicit none
@@ -190,6 +183,14 @@ module update
         type(extent), intent(in) :: dims
         real, dimension(-2:dims%imx+2,-2:dims%jmx+2,-2:dims%kmx+2,1:dims%n_var), intent(inout):: qp
         real, dimension(-2:dims%imx+2,-2:dims%jmx+2,-2:dims%kmx+2), intent(inout):: Temp
+        real, dimension(:, :, :, :), intent(inout)  :: residue
+        !< Store residue at each cell-center
+        real, dimension(:, :, :, :), intent(inout) :: F
+        !< Store fluxes throught the I faces
+        real, dimension(:, :, :, :), intent(inout) :: G
+        !< Store fluxes throught the J faces
+        real, dimension(:, :, :, :), intent(inout) :: H
+        !< Store fluxes throught the K faces
         real :: CFL 
         CFL = control%CFL
         !finding the updated Temperature using ideal gas law
@@ -197,66 +198,68 @@ module update
         Temp = qp(:,:,:,5)/(flow%R_gas*qp(:,:,:,1) )
         select case (trim(scheme%time_step_accuracy))
             case ("none")
-              call get_total_conservative_Residue(qp, Temp, control, scheme, flow, dims)
+              call get_total_conservative_Residue(qp, Temp, residue, F,G,H, control, scheme, flow, dims)
               call compute_time_step(qp, CFL, scheme, flow, dims) ! has to be after get_..._Residue()
[…]
