+++ src/face/flux/convective/slau.f90
@@ -12,16 +12,12 @@ module slau
     use global_vars, only : ynx, yny, ynz !face unit normal y
     use global_vars, only : znx, zny, znz !face unit normal z
     use global_vars, only : xA, yA, zA    !face area
-
-!    use global_vars, only : gm
-!    use global_vars, only : n_var
-!    use global_vars, only : turbulence
     use global_vars, only : process_id
     use global_vars, only : make_F_flux_zero
     use global_vars, only : make_G_flux_zero
     use global_vars, only : make_H_flux_zero
 
-    use utils, only: alloc, dealloc
+    use utils, only: alloc
     use face_interpolant, only: x_qp_left, x_qp_right 
     use face_interpolant, only: y_qp_left, y_qp_right
     use face_interpolant, only:  z_qp_left, z_qp_right
@@ -29,77 +25,80 @@ module slau
     implicit none
     private
 
-    real, public, dimension(:, :, :, :), allocatable, target :: F
-    !< Store fluxes throught the I faces
-    real, public, dimension(:, :, :, :), allocatable, target :: G
-    !< Store fluxes throught the J faces
-    real, public, dimension(:, :, :, :), allocatable, target :: H
-    !< Store fluxes throught the K faces
-    real, public, dimension(:, :, :, :), allocatable, target :: residue
-    !< Store residue at each cell-center
-    real, dimension(:, :, :, :), pointer :: flux_p
-    !< Pointer/alias for the either F, G, or H
-    integer :: imx, jmx, kmx, n_var
-
-    ! Public members
-    public :: setup_scheme
-    public :: destroy_scheme
+!    real, public, dimension(:, :, :, :), allocatable, target :: F
+!    !< Store fluxes throught the I faces
+!    real, public, dimension(:, :, :, :), allocatable, target :: G
+!    !< Store fluxes throught the J faces
+!    real, public, dimension(:, :, :, :), allocatable, target :: H
+!    !< Store fluxes throught the K faces
+!    real, public, dimension(:, :, :, :), allocatable, target :: residue
+!    !< Store residue at each cell-center
+!    real, dimension(:, :, :, :), pointer :: flux_p
+!    !< Pointer/alias for the either F, G, or H
+!    integer :: imx, jmx, kmx, n_var
+!
+!    ! Public members
+!    public :: setup_scheme
+!    public :: destroy_scheme
     public :: compute_fluxes
-    public :: get_residue
+!    public :: get_residue
     
     contains
 
-        subroutine setup_scheme(control, dims)
-          !< Allocate memory to the flux variables
+!        subroutine setup_scheme(control, dims)
+!          !< Allocate memory to the flux variables
+!
+!            implicit none
+!            type(controltype), intent(in) :: control
+!            type(extent), intent(in) :: dims
+!
+!            imx = dims%imx
+!            jmx = dims%jmx
+!            kmx = dims%kmx
+!
+!            n_var = control%n_var
+!
+!            DebugCall('setup_scheme')
+!
+!            call alloc(F, 1, imx, 1, jmx-1, 1, kmx-1, 1, n_var, &
+!                    errmsg='Error: Unable to allocate memory for ' // &
+!                        'F - slau.')
+!            call alloc(G, 1, imx-1, 1, jmx, 1, kmx-1, 1, n_var, &
+!                    errmsg='Error: Unable to allocate memory for ' // &
+!                        'G - slau.')
+!            call alloc(H, 1, imx-1, 1, jmx-1, 1, kmx, 1, n_var, &
+!                    errmsg='Error: Unable to allocate memory for ' // &
+!                        'H - slau.')
+!            call alloc(residue, 1, imx-1, 1, jmx-1, 1, kmx-1, 1, n_var, &
[…]
