### 2019-12-22 39e3ca4
Switch geometeries paramter: global to local
