+++ src/face/flux/convective/ausm.f90
@@ -7,16 +7,16 @@ module ausm
 #include "../../../debug.h"
 #include "../../../error.h"
      use vartypes
-    use global_vars, only : xnx, xny, xnz !face unit normal x
-    use global_vars, only : ynx, yny, ynz !face unit normal y
-    use global_vars, only : znx, zny, znz !face unit normal z
-    use global_vars, only : xA, yA, zA    !face area
+!    use global_vars, only : xnx, xny, xnz !face unit normal x
+!    use global_vars, only : ynx, yny, ynz !face unit normal y
+!    use global_vars, only : znx, zny, znz !face unit normal z
+!    use global_vars, only : xA, yA, zA    !face area
     use global_vars, only : process_id
     use global_vars, only : make_F_flux_zero
     use global_vars, only : make_G_flux_zero
     use global_vars, only : make_H_flux_zero
 
-    use utils, only: alloc
+!    use utils, only: alloc
     use face_interpolant, only: x_qp_left, x_qp_right 
     use face_interpolant, only: y_qp_left, y_qp_right
     use face_interpolant, only:  z_qp_left, z_qp_right
@@ -87,22 +87,24 @@ module ausm
 !
 !        end subroutine destroy_scheme
 
-        subroutine compute_flux(Flux, f_dir, flow, dims)
+        subroutine compute_flux(Flux, f_dir, faces, flags, flow, dims)
           !< A generalized subroutine to calculate
           !< flux through the input-argument direction, :x,y, or z
           !< which corresponds to the I,J, or K direction respectively
           !------------------------------------------------------------
 
             implicit none
+            integer, dimension(3), intent(in) :: flags
             type(extent), intent(in) :: dims
             type(flowtype), intent(in) :: flow
             real, dimension(:, :, :, :), intent(inout) :: Flux
             !< Store fluxes throught the any(I,J,K) faces
+            type(facetype), dimension(-2:dims%imx+2+flags(1),-2:dims%jmx+2+flags(2),-2:dims%kmx+2+flags(3)), intent(in) :: faces
             character, intent(in) :: f_dir
             !< Input direction for which flux are calcuated and store
             integer :: i, j, k 
             integer :: i_f, j_f, k_f ! Flags to determine face direction
-            real, dimension(:, :, :), pointer :: fA, nx, ny, nz
+            !real, dimension(:, :, :), pointer :: fA, nx, ny, nz
             real, dimension(:,:,:,:), pointer :: f_qp_left, f_qp_right
             real, dimension(1:dims%n_var) :: F_plus, F_minus
             real :: M_perp_left, M_perp_right
@@ -116,39 +118,42 @@ module ausm
             real :: temp_c
 
             DebugCall('compute_flux')
+            i_f = flags(1)
+            j_f = flags(2)
+            k_f = flags(3)
             
             select case (f_dir)
                 case ('x')
-                    i_f = 1
-                    j_f = 0
-                    k_f = 0
-                    !flux_p => F
-                    fA => xA
-                    nx => xnx
-                    ny => xny
-                    nz => xnz
+                    !i_f = 1
+                    !j_f = 0
+                    !k_f = 0
+                    !!flux_p => F
+                    !fA => xA
+                    !nx => xnx
+                    !ny => xny
+                    !nz => xnz
                     f_qp_left => x_qp_left
                     f_qp_right => x_qp_right
                 case ('y')
-                    i_f = 0
-                    j_f = 1
-                    k_f = 0
[…]
