+++ src/boundary/bc_primitive.f90
@@ -45,16 +45,16 @@ module bc_primitive
 !  use global_vars, only: kmx
 !  use global_vars, only: turbulence
 !  use global_vars, only: transition
-  use global_vars, only: xnx
-  use global_vars, only: xny
-  use global_vars, only: xnz
-  use global_vars, only: ynx
-  use global_vars, only: yny
-  use global_vars, only: ynz
-  use global_vars, only: znx
-  use global_vars, only: zny
-  use global_vars, only: znz
-  use global_vars, only: mu_t
+!  use global_vars, only: xnx
+!  use global_vars, only: xny
+!  use global_vars, only: xnz
+!  use global_vars, only: ynx
+!  use global_vars, only: yny
+!  use global_vars, only: ynz
+!  use global_vars, only: znx
+!  use global_vars, only: zny
+!  use global_vars, only: znz
+!  use global_vars, only: mu_t
 !  use global_vars, only: T_ref
   use global_vars, only: dist
   use global_vars, only: process_id
@@ -125,7 +125,7 @@ module bc_primitive
 
   contains
 
-    subroutine populate_ghost_primitive(state, control, scheme, flow, dims)
+    subroutine populate_ghost_primitive(state, Ifaces, Jfaces, Kfaces, control, scheme, flow, dims)
       !< Populate the state variables in the ghost cell
       !< with particular value based on the boundary conditio 
       !< being applied at that face
@@ -133,6 +133,12 @@ module bc_primitive
       type(extent), intent(in) :: dims
       real, dimension(-2:dims%imx+2, -2:dims%jmx+2, -2:dims%kmx+2, 1:dims%n_var), intent(inout), target :: state
       !< state variables
+      type(facetype), dimension(-2:dims%imx+3,-2:dims%jmx+2,-2:dims%kmx+2), intent(in) :: Ifaces
+      !< Input varaible which stores I faces' area and unit normal
+      type(facetype), dimension(-2:dims%imx+2,-2:dims%jmx+3,-2:dims%kmx+2), intent(in) :: Jfaces
+      !< Input varaible which stores J faces' area and unit normal
+      type(facetype), dimension(-2:dims%imx+2,-2:dims%jmx+2,-2:dims%kmx+3), intent(in) :: Kfaces
+      !< Input varaible which stores K faces' area and unit normal
       type(controltype), intent(in) :: control
       type(schemetype), intent(in) :: scheme
       type(flowtype), intent(in) :: flow
@@ -241,19 +247,19 @@ module bc_primitive
             call wall(face, dims)
 
           case(-6)
-            call slip_wall(face, dims)
+            call slip_wall(face, Ifaces, Jfaces, Kfaces, dims)
 
           case(-7)
             call pole(face, dims)
 
           case(-8)
-            call far_field(face, dims)
+            call far_field(face, Ifaces, Jfaces, Kfaces, dims)
 
           case(-9)
             call periodic_bc(face)
 
           case(-11)
-            call total_pressure(face, dims)
[…]
