+++ src/update.f90
@@ -27,7 +27,7 @@ module update
   use gradients  ,only :   gradw_x
   use gradients  ,only :   gradw_y
   use gradients  ,only :   gradw_z
-  use global_vars, only : volume
+!  use global_vars, only : volume
   use global_vars, only : dist
   use global_vars, only : mu
   use global_vars, only : mu_t
@@ -173,7 +173,7 @@ module update
 !      end subroutine destroy_update
 
 
-      subroutine get_next_solution(qp, Temp, residue, F,G,H, control, scheme, flow, dims)
+    subroutine get_next_solution(qp, Temp, residue, cells, F,G,H, Ifaces, Jfaces, Kfaces, control, scheme, flow, dims)
         !< Get solution at next time-step using scheme
         !< given in the input file.
         implicit none
@@ -191,6 +191,14 @@ module update
         !< Store fluxes throught the J faces
         real, dimension(:, :, :, :), intent(inout) :: H
         !< Store fluxes throught the K faces
+        type(celltype), dimension(-2:dims%imx+2,-2:dims%jmx+2,-2:dims%kmx+2), intent(in) :: cells
+        !< Input cell quantities: volume
+        type(facetype), dimension(-2:dims%imx+3,-2:dims%jmx+2,-2:dims%kmx+2), intent(in) :: Ifaces
+        !< Store face quantites for I faces 
+        type(facetype), dimension(-2:dims%imx+2,-2:dims%jmx+3,-2:dims%kmx+2), intent(in) :: Jfaces
+        !< Store face quantites for J faces 
+        type(facetype), dimension(-2:dims%imx+2,-2:dims%jmx+2,-2:dims%kmx+3), intent(in) :: Kfaces
+        !< Store face quantites for K faces 
         real :: CFL 
         CFL = control%CFL
         !finding the updated Temperature using ideal gas law
@@ -198,66 +206,68 @@ module update
         Temp = qp(:,:,:,5)/(flow%R_gas*qp(:,:,:,1) )
         select case (trim(scheme%time_step_accuracy))
             case ("none")
-              call get_total_conservative_Residue(qp, Temp, residue, F,G,H, control, scheme, flow, dims)
-              call compute_time_step(qp, CFL, scheme, flow, dims) ! has to be after get_..._Residue()
-              call update_with(qp, residue, scheme, flow, "conservative", 1. ,1., .FALSE.) 
+              call get_total_conservative_Residue(qp, Temp, cells, residue, F,G,H, Ifaces,Jfaces,Kfaces, control, scheme, flow, dims)
+              call compute_time_step(qp, CFL, cells, Ifaces, Jfaces, Kfaces, scheme, flow, dims) ! has to be after get_..._Residue()
+              call update_with(qp, residue, cells, scheme, flow, "conservative", 1. ,1., .FALSE.) 
             case ("RK4")
               R_store=0.
               U_store = qp
-              call get_total_conservative_Residue(qp, Temp, residue, F,G,H, control, scheme, flow, dims)
-              call compute_time_step(qp, CFL, scheme, flow, dims) ! has to be after get_..._Residue()
-              call update_with(qp, residue, scheme, flow, "conservative", 0.5  , 1., .FALSE., R_store, U_store) 
-              call get_total_conservative_Residue(qp, Temp, residue, F,G,H, control, scheme, flow, dims)
-              call update_with(qp, residue, scheme, flow, "conservative", 0.5  , 2., .FALSE., R_store, U_store) 
-              call get_total_conservative_Residue(qp, Temp, residue, F,G,H, control, scheme, flow, dims)
[…]

[…]
