### 2019-12-23 68f3508
Add: boundarytype
