+++ src/state.f90
@@ -29,179 +29,6 @@ module state
 
     contains
 
-!
-!        subroutine link_aliases(scheme)
-!          !< Setup state variable pointers
-!
-!            implicit none
-!            type(schemetype), intent(in) :: scheme
-!
-!            DebugCall("link_aliases")
-!
-!            density(-2:imx+2, -2:jmx+2, -2:kmx+2) => qp(:, :, :, 1)
-!            x_speed(-2:imx+2, -2:jmx+2, -2:kmx+2) => qp(:, :, :, 2)
-!            y_speed(-2:imx+2, -2:jmx+2, -2:kmx+2) => qp(:, :, :, 3)
-!            z_speed(-2:imx+2, -2:jmx+2, -2:kmx+2) => qp(:, :, :, 4)
-!            pressure(-2:imx+2, -2:jmx+2, -2:kmx+2) => qp(:, :, :, 5)
-!
-!
-!            select case (trim(scheme%turbulence))
-!
-!                case ("none")
-!                    !include nothing
-!                    continue
-!                
-!                case ("sst", "sst2003", "bsl", "des-sst", "kw")
-!                    tk(-2:imx+2, -2:jmx+2, -2:kmx+2) => qp(:, :, :, 6)
-!                    tw(-2:imx+2, -2:jmx+2, -2:kmx+2) => qp(:, :, :, 7)
-!
-!                case ("kkl")
-!                    tk(-2:imx+2, -2:jmx+2, -2:kmx+2) => qp(:, :, :, 6)
-!                    tkl(-2:imx+2, -2:jmx+2, -2:kmx+2) => qp(:, :, :, 7)
-!
-!                case ("sa", "saBC")
-!                    tv(-2:imx+2, -2:jmx+2, -2:kmx+2) => qp(:, :, :, 6)
-!
-!                case ("ke")
-!                    tk(-2:imx+2, -2:jmx+2, -2:kmx+2) => qp(:, :, :, 6)
-!                    te(-2:imx+2, -2:jmx+2, -2:kmx+2) => qp(:, :, :, 7)
-!
-!                case ("les")
-!                  continue
-!                  ! todo
-!
-!                case DEFAULT
-!                  Fatal_error
-!
-!            end select
-!
-!
-!            ! Transition modeling
-!            select case(trim(scheme%transition))
-!
-!              case('lctm2015')
-!                tgm(-2:imx+2, -2:jmx+2, -2:kmx+2) => qp(:, :, :, n_var)
-!!                tgm_inf => qp_inf(n_var)
-!
-!              case('bc', 'none')
-!                !do nothing
-!                continue
-!
-!              case DEFAULT
-!                Fatal_error
-!
-!            end Select
-!
-!        end subroutine link_aliases
-!
-
-
-!        subroutine unlink_aliases()
-!          !< Nullify the pointer link
-!
-!            implicit none
-!
-!            DebugCall("unlink_aliases")
-!
-!            nullify(density)
-!            nullify(x_speed)
-!            nullify(y_speed)
-!            nullify(z_speed)
-!            nullify(pressure)
-!
-!            nullify(density_inf)
-!            nullify(x_speed_inf)
-!            nullify(y_speed_inf)
-!            nullify(z_speed_inf)
-!            nullify(pressure_inf)
-!
-!            select case (trim(turbulence))
-!
-!                case ("none")
-!                    continue
-!
-!                  case ("sst", "sst2003", "bsl", "kw", "des-sst")
-!                    nullify(tk)
-!                    nullify(tw)
-!                    nullify(tk_inf)
-!                    nullify(tw_inf)
-!
-!                case ("kkl")
-!                    nullify(tk)
-!                    nullify(tkl)
-!                    nullify(tk_inf)
[…]
