### 2019-12-24 da8c80c
Fixed Bug: Relative residual scaling
