+++ src/viscous.f90
@@ -31,8 +31,8 @@ module viscous
   use gradients  , only : gradtgm_x
   use gradients  , only : gradtgm_y
   use gradients  , only : gradtgm_z
-  use global_vars, only : mu
-  use global_vars, only : mu_t
+  use viscosity, only : mu
+  use viscosity, only : mu_t
   use global_sst , only : sst_F1
   use global_sst , only : sigma_k1
   use global_sst , only : sigma_k2
