### 2019-12-25 251b741
Added: travis CI, a few unit tests
