+++ src/face/state/weno_NM.f90
@@ -9,112 +9,26 @@ module weno_NM
 #include "../../debug.h"
 
     use vartypes
-    use utils, only: alloc
-
     implicit none
     private
 
-    ! Private variables
-    real, dimension(:, :, :, :), allocatable, target :: x_qp_left
-      !< Store primitive state at the I-face left side
-    real, dimension(:, :, :, :), allocatable, target :: x_qp_right
-      !< Store primitive state at the I-face right side
-    real, dimension(:, :, :, :), allocatable, target :: y_qp_left
-      !< Store primitive state at the J-face left side
-    real, dimension(:, :, :, :), allocatable, target :: y_qp_right
-      !< Store primitive state at the J-face right side
-    real, dimension(:, :, :, :), allocatable, target :: z_qp_left
-      !< Store primitive state at the K-face left side
-    real, dimension(:, :, :, :), allocatable, target :: z_qp_right
-      !< Store primitive state at the K-face right side
-    
-    real, dimension(:, :, :, :), pointer :: f_qp_left
-    !< Generalized pointer for any I-J-K direction> f_qp_left can 
-    !< either point to x_qp_left, y_qp_left or z_qp_left
-    real, dimension(:, :, :, :), pointer :: f_qp_right
-    !< Generalized pointer for any I-J-K direction> f_qp_right can 
-    !< either point to x_qp_right, y_qp_right or z_qp_right
-
-    integer :: imx, jmx, kmx, n_var
     ! Public members
-    public :: setup_scheme
-!    public :: destroy_scheme
     public :: compute_weno_NM_states
-    public :: x_qp_left, x_qp_right
-    public :: y_qp_left, y_qp_right
-    public :: z_qp_left, z_qp_right
-
- !  TVD_scheme = trim('koren')
 
     contains
         
-        subroutine setup_scheme(control, dims)
-          !< Allocate memoery to all array which store state
-          !< the face.
 
-        implicit none
-        type(controltype), intent(in) :: control
-        type(extent), intent(in) :: dims
-
-        DebugCall('setup_weno_NM')
-
-        imx = dims%imx
-        jmx = dims%jmx
-        kmx = dims%kmx
-
-        n_var = control%n_var
-
-        call alloc(x_qp_left, 0, imx+1, 1, jmx-1, 1, kmx-1, 1, n_var, &
-            errmsg='Error: Unable to allocate memory for ' // &
-                'x_qp_left.')
-        call alloc(x_qp_right, 0, imx+1, 1, jmx-1, 1, kmx-1, 1, n_var, &
-            errmsg='Error: Unable to allocate memory for ' // &
-                'x_qp_right.')
-
-        call alloc(y_qp_left, 1, imx-1, 0, jmx+1, 1, kmx-1, 1, n_var, &
-            errmsg='Error: Unable to allocate memory for ' // &
-                'y_qp_left.')
-        call alloc(y_qp_right, 1, imx-1, 0, jmx+1, 1, kmx-1, 1, n_var, &
-            errmsg='Error: Unable to allocate memory for ' // &
-                'y_qp_right.')
-
-        call alloc(z_qp_left, 1, imx-1, 1, jmx-1, 0, kmx+1, 1, n_var, &
-            errmsg='Error: Unable to allocate memory for ' // &
-                'z_qp_left.')
-        call alloc(z_qp_right, 1, imx-1, 1, jmx-1, 0, kmx+1, 1, n_var, &
-            errmsg='Error: Unable to allocate memory for ' // &
-                'z_qp_right.')
-
-        end subroutine setup_scheme
-
-
-!        subroutine destroy_scheme()
-!          !< Deallocate all the array used 
-!
[…]
