### 2019-12-28 80ec53b
Added unit tests, updated docs
