+++ src/face/state/face_interpolant.f90
@@ -1,4 +1,6 @@
+ !< Interpolation of primitive variable from cell centers to faces
 module face_interpolant
+ !< Interpolation of primitive variable from cell centers to faces
 
 #include "../../debug.h"
 #include "../../error.h"
@@ -12,10 +14,14 @@ module face_interpolant
     private
 
 
-    real, dimension(:, :, :, :), allocatable, target :: x_qp_left, x_qp_right
-    real, dimension(:, :, :, :), allocatable, target :: y_qp_left, y_qp_right
-    real, dimension(:, :, :, :), allocatable, target :: z_qp_left, z_qp_right
-    real, dimension(:, :, :), allocatable :: pdif
+    real(wp), dimension(:, :, :, :), allocatable, target :: x_qp_left, x_qp_right
+    !< Store primitive state at the I-face 
+    real(wp), dimension(:, :, :, :), allocatable, target :: y_qp_left, y_qp_right
+    !< Store primitive state at the J-face 
+    real(wp), dimension(:, :, :, :), allocatable, target :: z_qp_left, z_qp_right
+    !< Store primitive state at the K-face 
+    real(wp), dimension(:, :, :), allocatable :: pdif
+    !< pressure difference
    !< Used for pressure based witch
 
 
@@ -29,37 +35,10 @@ module face_interpolant
 
     contains
 
-!        subroutine allocate_memory()
-!            implicit none
-!            call alloc(x_qp_left, 0, imx+1, 1, jmx-1, 1, kmx-1, 1, n_var, &
-!                    errmsg='Error: Unable to allocate memory for ' // &
-!                        'x_qp_left.')
-!            call alloc(x_qp_right, 0, imx+1, 1, jmx-1, 1, kmx-1, 1, n_var, &
-!                    errmsg='Error: Unable to allocate memory for ' // &
-!                        'x_qp_right.')
-!            call alloc(y_qp_left, 1, imx-1, 0, jmx+1, 1, kmx-1, 1, n_var, &
-!                    errmsg='Error: Unable to allocate memory for ' // &
-!                        'y_qp_left.')
-!            call alloc(y_qp_right, 1, imx-1, 0, jmx+1, 1, kmx-1, 1, n_var, &
-!                    errmsg='Error: Unable to allocate memory for ' // &
-!                        'y_qp_right.')
-!            call alloc(z_qp_left, 1, imx-1, 1, jmx-1, 0, kmx+1, 1, n_var, &
-!                    errmsg='Error: Unable to allocate memory for ' // &
-!                        'z_qp_left.')
-!            call alloc(z_qp_right, 1, imx-1, 1, jmx-1, 0, kmx+1, 1, n_var, &
-!                    errmsg='Error: Unable to allocate memory for ' // &
-!                        'z_qp_right.')
-!        call alloc(pdif, 0, imx, 0, jmx, 0, kmx, &
-!                errmsg='Error: Unable to allocate memory for' // &
-!                    'pdif')
-!        end subroutine allocate_memory
-
-
         subroutine setup_interpolant_scheme(dims)
             implicit none
-            !type(controltype), intent(in) :: control
-            !type(schemetype), intent(in) :: scheme
             type(extent), intent(in) :: dims
+            !< Extent of the domain:imx,jmx,kmx
 
             call alloc(x_qp_left, 0, dims%imx+1, 1, dims%jmx-1, 1, dims%kmx-1, 1, dims%n_var, &
                     errmsg='Error: Unable to allocate memory for x_qp_left.')
@@ -82,7 +61,9 @@ module face_interpolant
         subroutine extrapolate_cell_averages_to_faces(qp, dims)
             implicit none
[…]
