### 2019-12-28 5bb32e5
Corrected docs, unit tests
