+++ README.md
@@ -14,22 +14,32 @@ FEST3D (Finite-volume Explicit STructured 3-Dimensional) is computational fluid 
 [Full Documentation](https://fest3d.github.io/index.html) for the FEST-3D solver is generated with [FORD](https://github.com/Fortran-FOSS-Programmers/ford) documentation generator. Documentation is available at [website](https://fest3d.github.io/index.html) .
 
 ## Installation
-For installation instruction check out [Documentation](https://fest3d.github.io/page/01_install.html) guide.
+An install.sh bash script has been provided for ubuntu, CentOS and MacOS platforms. This script will take care of dependencies, build executable, run unit and integrated test and set `FEST3D` enviornment variable. You can run install.sh using the following command:
+```
+$ . install.sh
+```
+You might be prompted for password to install dependencies. This scirpt will add a following line at the end of either `.bashrc` or `.bash_profile` file
+```
+export FEST3D="<edit-path-to-FEST-3D-folder>/bin/FEST"
+```
+For manual installation instruction check out [Documentation](https://fest3d.github.io/page/01_install.html) guide.
 
 ### Dependencies
 FEST-3D has the following dependencies:
 
- * Fortran compiler
+ * Fortran and C++11 compiler
  * MPI library
  * CMAKE and MAKE
+ * Python3-numpy for integrated test
 
-So, all the dependencies can be installed using following commands:
+So, all the dependencies can be installed using following commands on ubuntu:
 ```
 $sudo apt-get update
 $sudo apt-get install build-essential
 $sudo apt-get install cmake
 $sudo apt-get install gfortran
 $sudo apt-get install mpich
+$sudo apt-get install python3-numpy
 ```
 The CMAKE implementation in FEST-3D is adapted from the template developed by [Seth Morton](https://github.com/SethMMorton) available at [GitHub](https://github.com/SethMMorton/cmake_fortran_template).
 
@@ -39,67 +49,24 @@ To build the FEST-3D code, we first generate a new "build" directory/folder in t
 Following list of comnands will build the executalbe
 ```
 $mkdir build && cd build
-$FC=mpif90 cmake ..
+$cmake ..
 $make -j 4
 ```
 These commands will create a binary file named FEST3D in the ```bin``` folder; located in the root directory of the FEST-3D code.
 
+### Check
+Make sure the `FEST3D` enviornment variable is set properly by using following command:
+```
+echo $FEST3D
+```
+If output is string of zero length then execute following and additionaly add it to `.bashrc` or `.bash_profile` for permanent effect.
+```
+export FEST3D="<edit-path-to-FEST-3D-folder>/bin/FEST"
+```
+
 
 ## Tutorials
-For tutorials check out description of [test cases](https://fest3d.github.io/page/05_tutorials/index.html).
-
-## Expected Inputs
-A python script is provided in the run folder to ease the user interface with FEST-3D. Following inputs are
-required in order to setup any case in FEST-3D.   
-
-| Variable |  Expected Input | Description|
-|:---------|:----------------|:----------------------|
-|RunDir    |String           | Give any name to the Run directory |
[…]
