+++ tests/files/bc_00.md
@@ -1,0 +1,34 @@
+BOUNDARY CONDITIONS CONFIGURATION
+=================================
+
+# imn
+- COPY_DENSITY
+- COPY_X_SPEED
+- COPY_Y_SPEED
+- COPY_X_SPEED
+- FIX_PRESSURE 100000
+
+# imx
+- INTERFACE
+
+# jmn
+- COPY_DENSITY
+- COPY_PRESSURE
+- FLOW_TANGENCY
+
+# jmx
+- COPY_DENSITY
+- COPY_PRESSURE
+- FLOW_TANGENCY
+
+# kmn
+- COPY_DENSITY
+- COPY_PRESSURE
+- FLOW_TANGENCY
+
+# kmx
+- COPY_DENSITY
+- COPY_PRESSURE
+- FLOW_TANGENCY
+
+FIN
