+++ tests/files/fvscheme.md
@@ -1,0 +1,30 @@
+
+Scheme File
+===========
+## Inviscid Flux Scheme
+ausm
+
+## Higher Order Method
+muscl
+
+## Switch: Limiter - Pressure based switch
+0 0 0  0 0 0
+
+## Turbulence Limiter Switch
+1 1 1
+
+## Turbulence model
+sst
+
+## Transition model
+none
+
+## Time Step
+l
+
+## Time Integration Method
+implicit
+
+## Higher Order Boundary Conditions
+0
+
