+++ tests/files/control.md
@@ -1,0 +1,42 @@
+
+Control File
+===========
+## CFL
+100
+
+## Restart level
+2
+
+## Maximum Iterations
+10000
+
+## Save Iterations
+10000
+
+## Output File format
+tecplot
+
+## Output Data format
+ASCII
+
+## Input File format
+tecplot
+
+## Input Data format
+ASCII
+
+## Write Precision
+6
+
+## Purge Write
+1
+
+## Residual write Interval
+100
+
+## Tolerance
+1e-13 Continuity_abs
+
+## Debug Level
+5
+

[…]
