### 2020-01-05 5d677a9
Added: Unit tests, Travis CI, Install script, and Derived data types
* add grid derived data type

* Add: CTest, geomerty derived data type

* Add: controltype for control file

* ADD: schemetype and flowtype

* Add: read and write variable to controltype

* Switched: global state variables to local

* Switched->Flux vectors: from global to local

* Switch geometeries paramter: global to local

* Add: boundarytype

* Fixed Bug: Relative residual scaling

* Added: travis CI, a few unit tests

* Added unit tests, updated docs

* Corrected docs, unit tests
