+++ src/viscous.f90
@@ -1,0 +1,757 @@
+  !< The viscous module contains the viscous fluxes calculations 
+module viscous
+  !< The viscous module contains the viscous fluxes calculations
+  !-----------------------------------------------------------------
+#include "error.inc"
+  use vartypes
+  use gradients  , only : gradu_x
+  use gradients  , only : gradu_y
+  use gradients  , only : gradu_z
+  use gradients  , only : gradv_x
+  use gradients  , only : gradv_y
+  use gradients  , only : gradv_z
+  use gradients  , only : gradw_x
+  use gradients  , only : gradw_y
+  use gradients  , only : gradw_z
+  use gradients  , only : gradT_x
+  use gradients  , only : gradT_y
+  use gradients  , only : gradT_z
+  use gradients  , only : gradtk_x
+  use gradients  , only : gradtk_y
+  use gradients  , only : gradtk_z
+  use gradients  , only : gradtw_x
+  use gradients  , only : gradtw_y
+  use gradients  , only : gradtw_z
+  use gradients  , only : gradtkl_x
+  use gradients  , only : gradtkl_y
+  use gradients  , only : gradtkl_z
+  use gradients  , only : gradtv_x
+  use gradients  , only : gradtv_y
+  use gradients  , only : gradtv_z
+  use gradients  , only : gradtgm_x
+  use gradients  , only : gradtgm_y
+  use gradients  , only : gradtgm_z
+  use viscosity, only : mu
+  use viscosity, only : mu_t
+  use global_sst , only : sst_F1
+  use global_sst , only : sigma_k1
+  use global_sst , only : sigma_k2
+  use global_sst , only : sigma_w1
+  use global_sst , only : sigma_w2
+  use global_kkl , only : sigma_k
+  use global_kkl , only : sigma_phi
+  use global_sa  , only : sigma_sa
+  use global_sa  , only : cb2
+  use utils      , only : alloc
+  implicit none
+  private
+
+  integer :: imx, jmx, kmx
+
+  public :: compute_viscous_fluxes
+
+  contains
+
+    subroutine compute_viscous_fluxes(F, G, H, qp, cells, Ifaces, Jfaces, Kfaces, scheme, flow, dims)
+      !< Call to all viscous flux subroutine based on 
+      !< the drection and turbulence/transition model being
+      !< used
+
+        implicit none
+        type(schemetype), intent(in) :: scheme
+        !< finite-volume Schemes
+        type(flowtype), intent(in) :: flow
+        !< Information about fluid flow: freestream-speed, ref-viscosity,etc.
+        type(extent), intent(in) :: dims
+        !< Extent of the domain:imx,jmx,kmx
+        real(wp), dimension(-2:dims%imx+2, -2:dims%jmx+2, -2:dims%kmx+2, 1:dims%n_var), intent(in) :: qp
+        real(wp), dimension(:, :, :, :), intent(inout) :: F, G, H
+        type(celltype), dimension(-2:dims%imx+2,-2:dims%jmx+2,-2:dims%kmx+2), intent(in) :: cells
+        !< Input cell quantities: volume
+        type(facetype), dimension(-2:dims%imx+3,-2:dims%jmx+2,-2:dims%kmx+2), intent(in) :: Ifaces
+        !< Store face quantites for I faces 
[…]
