### 2020-01-05 #1555 @jayten
@labarba
The proposed work has been completed. The changes have been tested and merged with the master branch. Please review the changes [here](https://github.com/FEST3D/FEST-3D/pull/10).

The summary of the changes made is:

1. Added **install bash script** for automatic installation of FEST-3D with dependencies on Ubuntu, CentOS, and MacOS platforms. This script will install dependencies, build executable, run unit, and integrated test.
2. Added **Travis CI**.
3. Since **Python2** will no longer be maintained past 2020, Python script has been moved from version 2 to **version 3**
4. Clustering similar variables using **derived data-types**, which helps in making the argument list of various subroutines shorter.
5. Instead of using global variables, most variables are passed as subroutine arguments.
6. Updated CMake to automatically include for MPICH library, which removes the need to specify Fortran compiler while executing CMake file.
7. Integrated **Ctest** for Fortran with CMake to avoid additional dependencies for unit test setup.
8. Added 25 **unit tests**.
9. Removed commented code.
10. Deleted unnecessary archived/unused files.
11. Update documentation with the above changes.

Additionally, the install script also sets the "FEST3D" environment variable which will avoid the error faced by both reviewers while setting up the test cases using Python script in the "run/" directory.
