+++ src/read_write/read/read_output_tec.f90
@@ -81,10 +81,6 @@ module read_output_tec
               tk(-2:imx+2, -2:jmx+2, -2:kmx+2) => state(:, :, :, 6)
               te(-2:imx+2, -2:jmx+2, -2:kmx+2) => state(:, :, :, 7)
 
-          case ("les")
-            continue
-            ! todo
-
           case DEFAULT
             Fatal_error
       end select
@@ -93,7 +89,6 @@ module read_output_tec
       select case(trim(scheme%transition))
         case('lctm2015')
           tgm(-2:imx+2, -2:jmx+2, -2:kmx+2) => state(:, :, :, 8)
-!          tgm_inf => qp_inf(n_var)
 
         case('bc', 'none')
           !do nothing
