### 2020-02-03 #1555 @rouson
@labarba  This is a very impressive step forward.  I recommend acceptance.  I just installed FEST-3D on macOS and had a much better experience with the process this time.  After I issued the one simple installation command provided in the repository's README.md, the installation completed without problem, including running and passing all 25 unit tests and all 3 integration tests.  Ease of installation greatly improves the user experience and passing tests greatly improve user confidence.

I also looked at several source files and believe the code at least makes better use of Fortran 90.  I encourage the authors to explore and exploit the numerous major advances made in more recent Fortran standards, including support for object-oriented programming, parallel programming, and functional programming.  I suspect the code could still be made quite a bit less verbose through the use of array statements and more robust by either eliminating pointers or incorporating reference counting.  At a minimum, all pointers should be default-initialized to `null()` so that no pointer starts life in an undefined state.  I hope to find the time to make more suggestions via pull request.
