### 2020-02-04 #1555 @hariradh
I confirm that I recommend acceptance.

> On 3 Feb 2020, at 3:01 PM, Lorena A. Barba <notifications@github.com> wrote:
>
> @hariradh <https://github.com/hariradh> — I see all your checklist items are ticked off. Can we have a confirmation here that you recommend acceptance?
>
