### 2020-02-09 #1555 @labarba
Please merge https://github.com/FEST3D/FEST-3D/pull/11 which provides small typo fixes.

Then, you need to do these final steps:

 [ ] Please make a tagged release of your software, and list the version tag of the archived version here.
- [ ] Archive the reviewed software in Zenodo (or a similar service)
- [ ] Please list the Zenodo DOI of the archived version here.

You may find this helpful: https://guides.github.com/activities/citable-code/
