### 2020-02-09 #1555 @jayten
I have released the reviewed FEST-3D software as a new version with tag [v1.1](https://github.com/FEST3D/FEST-3D/releases/tag/v1.1) and archived the same on Zenodo with DOI: [10.5281/zenodo.3660408 ](https://zenodo.org/record/3660408).
