### 2020-02-10 #1555 @labarba
👋 @jayten — Please go over your reference list and add DOIs where available. Use the `doi` field and put just the identifier, not the full URL. Also, check the capitalization of journal titles, and acronyms like AIAA (appears as Aiaa in places). If you need to protect caps, use`{}`.
