### 2020-02-10 #1555 @jayten
I have added `doi` field for each bib entry and also removed the unused entries from the paper.bib file. Additionally, I included {} in the `booktitle` field to make sure that AIAA acroynms appears correctly.
