### 2020-02-10 #1555 @labarba
Congratulations, @jayten, your JOSS article is published! 🚀

Huge thanks to our reviewers: @hariradh, @rouson — we truly appreciate your contribution to this adventure in scholarly publishing 🙏
