### 2020-02-10 #1555 @jayten
On behalf of the FEST-3D team, I would like to thank @hariradh and @rouson for their time and effort to help us improve our work.

Also, a big thanks to @whedon and @labarba, for providing such a unique platform for publishing our work.
